Update clicfg.go
Fixes in clicfg for  https://github.com/neo4j/aura-cli/issues/108

+++ common/clicfg/clicfg.go
@@ -46,6 +46,9 @@ func NewConfig(fs afero.Fs, version string) *Config {
 	bindEnvironmentVariables(Viper)
 	setDefaultValues(Viper)
 
+	// This ensures we have a config file - See issue https://github.com/neo4j/aura-cli/issues/108 
+	Viper.SafeWriteConfig()
+
 	if err := Viper.ReadInConfig(); err != nil {
 		if _, ok := err.(viper.ConfigFileNotFoundError); ok {
 			if err := fs.MkdirAll(configPath, 0755); err != nil {
@@ -109,24 +112,37 @@ func (config *AuraConfig) Get(key string) interface{} {
 }
 
 func (config *AuraConfig) Set(key string, value string) {
+
 	filename := config.viper.ConfigFileUsed()
-	data := fileutils.ReadFileSafe(config.fs, filename)
 
-	updateConfig, err := sjson.Set(string(data), fmt.Sprintf("aura.%s", key), value)
-	if err != nil {
-		panic(err)
-	}
+	// We should always have a config file by this point but just in case
+	// Does the config file exist before we try and write to it?
+	// See  https://github.com/neo4j/aura-cli/issues/108 for why we should do this
+	if fileutils.FileExists(config.fs, filename) {
 
-	updatedAuraBaseUrl := config.auraBaseUrlOnBetaEnabledChange(key, value)
-	if updatedAuraBaseUrl != "" {
-		intermediateUpdateConfig, err := sjson.Set(string(updateConfig), "aura.base-url", updatedAuraBaseUrl)
+		data := fileutils.ReadFileSafe(config.fs, filename)
+
+		updateConfig, err := sjson.Set(string(data), fmt.Sprintf("aura.%s", key), value)
 		if err != nil {
 			panic(err)
 		}
-		updateConfig = intermediateUpdateConfig
+
+		updatedAuraBaseUrl := config.auraBaseUrlOnBetaEnabledChange(key, value)
+		if updatedAuraBaseUrl != "" {
+			intermediateUpdateConfig, err := sjson.Set(string(updateConfig), "aura.base-url", updatedAuraBaseUrl)
+			if err != nil {
+				panic(err)
+			}
+			updateConfig = intermediateUpdateConfig
+		}
+
+		fileutils.WriteFile(config.fs, filename, []byte(updateConfig))
+
+	} else {
+	        err := fmt.Errorf("unable to write to config file")
+		panic(err)
 	}
 
-	fileutils.WriteFile(config.fs, filename, []byte(updateConfig))
 }
 
 func (config *AuraConfig) Print(cmd *cobra.Command) {
